refactor(qprog): split optimizers into a package and harden selection
Promote optimizers.py into a package — one module per optimizer plus a
shared _linalg helper — and reorganize the optimizer tests to match,
with a two-tier contract suite. Guard best-result selection against
non-finite losses across Scipy, Pymoo, MonteCarlo, and GridSearch, and
return MonteCarlo's best-ever iterate instead of the last population's.
