--- tests/qprog/test_optimizers.py
@@ -1,1241 +1,0 @@
-# SPDX-FileCopyrightText: 2025-2026 Qoro Quantum Ltd <divi@qoroquantum.de>
-#
-# SPDX-License-Identifier: Apache-2.0
-
-import warnings
-
-import numpy as np
-import pytest
-from scipy.optimize import OptimizeResult
-
-from divi.qprog.optimizers import (
-    GridSearchOptimizer,
-    MonteCarloOptimizer,
-    Optimizer,
-    PymooMethod,
-    PymooOptimizer,
-    ScipyMethod,
-    ScipyOptimizer,
-)
-from tests.qprog._optimizer_contracts import (
-    verify_load_state_raises_file_not_found,
-    verify_save_creates_checkpoint_file,
-    verify_save_creates_directory_if_needed,
-    verify_save_load_round_trip,
-    verify_save_without_prior_run_raises,
-)
-
-
-def sphere_cost_fn_population(params: np.ndarray) -> np.ndarray:
-    """Sphere cost function (sum of squares) for a population of parameter sets."""
-    if params.ndim != 2:
-        raise ValueError(
-            "Input params for population cost function must be a 2D array."
-        )
-    return np.sum(params**2, axis=1)
-
-
-def sphere_cost_fn_single(params: np.ndarray) -> float:
-    """Sphere cost function (sum of squares) for a single parameter set."""
-    if params.ndim != 1:
-        # Allow single-row 2D array for convenience with some optimizers
-        if params.ndim == 2 and params.shape[0] == 1:
-            params = params.squeeze(0)
-        else:
-            raise ValueError(
-                "Input params for single cost function must be a 1D array."
-            )
-    return float(np.sum(params**2))
-
-
-# Locally extends the conftest variants with a MonteCarlo "keep best" mode that
-# only the optimizer-contract suite exercises.
-_CONTRACT_OPTIMIZER_VARIANTS = [
-    ("monte-carlo", lambda: MonteCarloOptimizer(population_size=5, n_best_sets=2)),
-    ("l-bfgs-b", lambda: ScipyOptimizer(method=ScipyMethod.L_BFGS_B)),
-    ("cobyla", lambda: ScipyOptimizer(method=ScipyMethod.COBYLA)),
-    ("nelder-mead", lambda: ScipyOptimizer(method=ScipyMethod.NELDER_MEAD)),
-    ("cmaes", lambda: PymooOptimizer(method=PymooMethod.CMAES, population_size=10)),
-    ("de", lambda: PymooOptimizer(method=PymooMethod.DE, population_size=5)),
-    (
-        "monte-carlo-keep-best",
-        lambda: MonteCarloOptimizer(
-            population_size=10, n_best_sets=3, keep_best_params=True
-        ),
-    ),
-]
-
-
-@pytest.fixture(
-    params=_CONTRACT_OPTIMIZER_VARIANTS,
-    ids=lambda param: param[0],
-)
-def contract_optimizer(request):
-    """Optimizer-contract variants (extends the conftest set with ``monte-carlo-keep-best``).
-
-    Named distinctly from the conftest-level ``optimizer`` fixture to avoid a
-    silent shadow — the contract suite needs the extra variant that the
-    cross-suite conftest fixture does not provide.
-    """
-    _, factory = request.param
-    return factory()
-
-
-class TestOptimizerContract:
-    """A contract of tests that every optimizer should pass."""
-
-    @pytest.fixture(autouse=True)
-    def setup(self):
-        self.n_params = 4
-        self.rng = np.random.default_rng(42)
-
-    def _get_initial_params(self, optimizer: Optimizer) -> np.ndarray:
-        """Generate initial parameters with the correct shape for the optimizer."""
-        shape = (optimizer.n_param_sets, self.n_params)
-        return self.rng.random(shape) * 2 * np.pi
-
-    def test_final_result_is_consistent(self, contract_optimizer: Optimizer):
-        """Verify that the final returned parameters match the final cost."""
-        initial_params = self._get_initial_params(contract_optimizer)
-        cost_fn = (
-            sphere_cost_fn_single
-            if contract_optimizer.n_param_sets == 1
-            else sphere_cost_fn_population
-        )
-
-        result = contract_optimizer.optimize(cost_fn, initial_params, max_iterations=5)
-
-        assert isinstance(result, OptimizeResult)
-        assert result.x.shape == (self.n_params,)
-        recalculated_fun = sphere_cost_fn_single(result.x)
-        assert np.isclose(
-            result.fun, recalculated_fun
-        ), "Final cost value should correspond to the final parameters."
-
-    def test_supports_checkpointing_matches_get_config(
-        self, contract_optimizer: Optimizer
-    ):
-        """``supports_checkpointing`` must agree with whether the optimizer can
-        actually serialize its config (get_config raises for those that cannot)."""
-        if contract_optimizer.supports_checkpointing:
-            assert isinstance(contract_optimizer.get_config(), dict)
-        else:
-            with pytest.raises(NotImplementedError):
-                contract_optimizer.get_config()
-
-    def test_callback_provides_consistent_results(self, contract_optimizer: Optimizer):
-        """Verify that parameters and costs are consistent in each callback call."""
-        initial_params = self._get_initial_params(contract_optimizer)
-        cost_fn = (
-            sphere_cost_fn_single
-            if contract_optimizer.n_param_sets == 1
-            else sphere_cost_fn_population
-        )
-
-        callback_results = []
-
-        def callback(intermediate_result: OptimizeResult):
-            callback_results.append(intermediate_result)
-
-        contract_optimizer.optimize(
-            cost_fn, initial_params, callback_fn=callback, max_iterations=5
-        )
-
-        assert len(callback_results) > 0
-        for res in callback_results:
-            assert isinstance(res, OptimizeResult)
-            assert res.x.shape == (contract_optimizer.n_param_sets, self.n_params)
-            assert res.fun.shape == (contract_optimizer.n_param_sets,)
-
-            if contract_optimizer.n_param_sets == 1:
-                recalculated_fun = sphere_cost_fn_single(res.x)
-            else:
-                recalculated_fun = sphere_cost_fn_population(res.x)
-
-            # L-BFGS-B can provide a stale loss value in its callback due to its
-            # line-search mechanism. For this optimizer only, we relax the test
-            # to confirm the loss is a valid number, but not that it's perfectly
-            # in sync with the parameters, as that would require re-computation.
-            if (
-                isinstance(contract_optimizer, ScipyOptimizer)
-                and contract_optimizer.method == ScipyMethod.L_BFGS_B
-            ):
-                assert np.isfinite(res.fun).all()
-            else:
-                assert np.allclose(res.fun, recalculated_fun)
-
-    def test_reset_allows_reusing_optimizer(self, contract_optimizer: Optimizer):
-        """Verify that reset() allows reusing an optimizer for fresh optimization runs."""
-        initial_params = self._get_initial_params(contract_optimizer)
-        cost_fn = (
-            sphere_cost_fn_single
-            if contract_optimizer.n_param_sets == 1
-            else sphere_cost_fn_population
-        )
-
-        # Run first optimization
-        result1 = contract_optimizer.optimize(cost_fn, initial_params, max_iterations=3)
-        assert isinstance(result1, OptimizeResult)
-
-        # Reset and run second optimization with different initial params
-        contract_optimizer.reset()
-        different_initial_params = self._get_initial_params(contract_optimizer)
-        result2 = contract_optimizer.optimize(
-            cost_fn, different_initial_params, max_iterations=3
-        )
-        assert isinstance(result2, OptimizeResult)
-
-        # Both results should be valid
-        assert result1.x.shape == (self.n_params,)
-        assert result2.x.shape == (self.n_params,)
-        assert np.isfinite(result1.fun)
-        assert np.isfinite(result2.fun)
-
-    def test_reset_preserves_optimizer_configuration(
-        self, contract_optimizer: Optimizer
-    ):
-        """Verify that reset() does not affect optimizer configuration."""
-        # Store original configuration
-        if isinstance(contract_optimizer, ScipyOptimizer):
-            original_method = contract_optimizer.method
-        elif isinstance(contract_optimizer, PymooOptimizer):
-            original_method = contract_optimizer.method
-            original_pop_size = contract_optimizer.population_size
-        elif isinstance(contract_optimizer, MonteCarloOptimizer):
-            original_pop_size = contract_optimizer.population_size
-            original_n_best = contract_optimizer.n_best_sets
-            original_keep_best = contract_optimizer.keep_best_params
-
-        # Run optimization to set state
-        initial_params = self._get_initial_params(contract_optimizer)
-        cost_fn = (
-            sphere_cost_fn_single
-            if contract_optimizer.n_param_sets == 1
-            else sphere_cost_fn_population
-        )
-        contract_optimizer.optimize(cost_fn, initial_params, max_iterations=2)
-
-        # Reset and verify configuration is unchanged
-        contract_optimizer.reset()
-
-        if isinstance(contract_optimizer, ScipyOptimizer):
-            assert contract_optimizer.method == original_method
-        elif isinstance(contract_optimizer, PymooOptimizer):
-            assert contract_optimizer.method == original_method
-            assert contract_optimizer.population_size == original_pop_size
-        elif isinstance(contract_optimizer, MonteCarloOptimizer):
-            assert contract_optimizer.population_size == original_pop_size
-            assert contract_optimizer.n_best_sets == original_n_best
-            assert contract_optimizer.keep_best_params == original_keep_best
-
-    def test_reset_can_be_called_multiple_times(self, contract_optimizer: Optimizer):
-        """Verify that reset() can be called multiple times safely."""
-        initial_params = self._get_initial_params(contract_optimizer)
-        cost_fn = (
-            sphere_cost_fn_single
-            if contract_optimizer.n_param_sets == 1
-            else sphere_cost_fn_population
-        )
-
-        # Run optimization
-        contract_optimizer.optimize(cost_fn, initial_params, max_iterations=2)
-
-        # Call reset multiple times
-        contract_optimizer.reset()
-        contract_optimizer.reset()
-        contract_optimizer.reset()
-
-        # Should still be able to optimize after multiple resets
-        result = contract_optimizer.optimize(cost_fn, initial_params, max_iterations=2)
-        assert isinstance(result, OptimizeResult)
-        assert result.x.shape == (self.n_params,)
-
-
-class TestMonteCarloOptimizer:
-    """Specific tests for MonteCarloOptimizer features."""
-
-    @pytest.fixture(autouse=True)
-    def setup(self):
-        self.n_params = 4
-        self.rng = np.random.default_rng(42)
-        self.population_size = 10
-        self.n_best_sets = 3
-
-    def _create_optimizer(self, keep_best_params: bool) -> MonteCarloOptimizer:
-        """Helper to create a MonteCarloOptimizer with standard test parameters."""
-        return MonteCarloOptimizer(
-            population_size=self.population_size,
-            n_best_sets=self.n_best_sets,
-            keep_best_params=keep_best_params,
-        )
-
-    def _create_initial_params(self) -> np.ndarray:
-        """Helper to create initial parameters with correct shape."""
-        return self.rng.random((self.population_size, self.n_params)) * 2 * np.pi
-
-    def _get_best_params(self, population: np.ndarray, n_best: int) -> np.ndarray:
-        """Extract the best n_best parameter sets from a population."""
-        losses = sphere_cost_fn_population(population)
-        best_indices = np.argpartition(losses, n_best - 1)[:n_best]
-        return population[best_indices]
-
-    def _run_optimization_with_callback(
-        self,
-        optimizer: MonteCarloOptimizer,
-        initial_params: np.ndarray,
-        max_iterations: int = 3,
-    ) -> list[np.ndarray]:
-        """Run optimization and collect all populations via callback."""
-        all_populations = []
-
-        def callback(intermediate_result: OptimizeResult):
-            all_populations.append(intermediate_result.x.copy())
-
-        optimizer.optimize(
-            sphere_cost_fn_population,
-            initial_params,
-            callback_fn=callback,
-            max_iterations=max_iterations,
-            rng=self.rng,
-        )
-        return all_populations
-
-    def test_keep_best_params_validation_and_property(self):
-        """Test validation errors and property access."""
-        # Test validation error
-        with pytest.raises(
-            ValueError,
-            match="If keep_best_params is True, n_best_sets must be less than population_size.",
-        ):
-            MonteCarloOptimizer(
-                population_size=10, n_best_sets=10, keep_best_params=True
-            )
-
-        # Test no error when keep_best_params=False
-        optimizer_no_error = MonteCarloOptimizer(
-            population_size=10, n_best_sets=10, keep_best_params=False
-        )
-        assert optimizer_no_error.keep_best_params is False
-
-        # Test property returns correct values
-        optimizer_false = self._create_optimizer(keep_best_params=False)
-        optimizer_true = self._create_optimizer(keep_best_params=True)
-        assert optimizer_false.keep_best_params is False
-        assert optimizer_true.keep_best_params is True
-
-    @pytest.mark.parametrize("keep_best_params", [True, False])
-    def test_keep_best_params_behavior(self, keep_best_params: bool):
-        """Test that keep_best_params correctly controls whether best parameters are kept."""
-        optimizer = self._create_optimizer(keep_best_params=keep_best_params)
-        initial_params = self._create_initial_params()
-
-        all_populations = self._run_optimization_with_callback(
-            optimizer, initial_params, max_iterations=3
-        )
-
-        assert len(all_populations) >= 2
-
-        # Get best params from initial population
-        best_params_first = self._get_best_params(
-            all_populations[0], optimizer.n_best_sets
-        )
-
-        # Check second population
-        second_population = all_populations[1]
-        assert second_population.shape == (optimizer.population_size, self.n_params)
-        first_n_best_in_second = second_population[: optimizer.n_best_sets]
-
-        # Check if each of the best parameters from the previous generation is present
-        # in the best parameters of the new generation.
-        is_present = [
-            np.any(np.all(np.isclose(first_n_best_in_second, p), axis=1))
-            for p in best_params_first
-        ]
-
-        if keep_best_params:
-            # If we keep the best params, all of them should be present.
-            assert all(
-                is_present
-            ), "Not all best parameters were kept when keep_best_params=True"
-        else:
-            # If we don't, none of them should be present as exact copies.
-            assert not any(
-                is_present
-            ), "A best parameter was kept as an exact copy when keep_best_params=False"
-
-    def test_reset_clears_monte_carlo_state(self):
-        """Test that reset() clears all internal state variables for MonteCarloOptimizer."""
-        optimizer = self._create_optimizer(keep_best_params=False)
-        initial_params = self._create_initial_params()
-
-        # Run optimization to set state
-        optimizer.optimize(
-            sphere_cost_fn_population, initial_params, max_iterations=3, rng=self.rng
-        )
-
-        # Verify state is set
-        assert optimizer._curr_population is not None
-        assert optimizer._curr_evaluated_population is not None
-        assert optimizer._curr_losses is not None
-        assert optimizer._curr_iteration is not None
-        assert optimizer._curr_rng_state is not None
-
-        # Reset and verify state is cleared
-        optimizer.reset()
-        assert optimizer._curr_population is None
-        assert optimizer._curr_evaluated_population is None
-        assert optimizer._curr_losses is None
-        assert optimizer._curr_iteration is None
-        assert optimizer._curr_rng_state is None
-
-    def test_save_state_creates_checkpoint_file(self, tmp_path):
-        """Test that save_state() creates the expected checkpoint file."""
-        optimizer = self._create_optimizer(keep_best_params=False)
-        initial_params = self._create_initial_params()
-        verify_save_creates_checkpoint_file(
-            optimizer,
-            initial_params,
-            sphere_cost_fn_population,
-            self.rng,
-            tmp_path,
-            MonteCarloOptimizer.load_state,
-        )
-
-    def test_save_state_preserves_configuration(self, tmp_path):
-        """Test that save_state() saves optimizer configuration correctly."""
-        optimizer = self._create_optimizer(keep_best_params=True)
-        initial_params = self._create_initial_params()
-
-        # Run optimization to set state
-        optimizer.optimize(
-            sphere_cost_fn_population, initial_params, max_iterations=2, rng=self.rng
-        )
-
-        checkpoint_dir = str(tmp_path / "checkpoint")
-        optimizer.save_state(checkpoint_dir)
-
-        # Load and verify configuration
-        loaded_optimizer = MonteCarloOptimizer.load_state(checkpoint_dir)
-        assert loaded_optimizer.population_size == optimizer.population_size
-        assert loaded_optimizer.n_best_sets == optimizer.n_best_sets
-        assert loaded_optimizer.keep_best_params == optimizer.keep_best_params
-
-    def test_load_state_restores_optimization_state(self, tmp_path):
-        """Test that load_state() correctly restores all optimization state."""
-        optimizer = self._create_optimizer(keep_best_params=False)
-        initial_params = self._create_initial_params()
-
-        # Run optimization to set state
-        optimizer.optimize(
-            sphere_cost_fn_population, initial_params, max_iterations=3, rng=self.rng
-        )
-
-        # Capture state before saving
-        original_population = optimizer._curr_population.copy()
-        original_evaluated = optimizer._curr_evaluated_population.copy()
-        original_losses = optimizer._curr_losses.copy()
-        original_iteration = optimizer._curr_iteration
-
-        checkpoint_dir = str(tmp_path / "checkpoint")
-        optimizer.save_state(checkpoint_dir)
-
-        # Load state
-        loaded_optimizer = MonteCarloOptimizer.load_state(checkpoint_dir)
-
-        # Verify all state is restored
-        np.testing.assert_array_equal(
-            loaded_optimizer._curr_population, original_population
-        )
-        np.testing.assert_array_equal(
-            loaded_optimizer._curr_evaluated_population, original_evaluated
-        )
-        np.testing.assert_array_equal(loaded_optimizer._curr_losses, original_losses)
-        assert loaded_optimizer._curr_iteration == original_iteration
-        assert loaded_optimizer._curr_rng_state is not None
-
-    def test_save_load_round_trip(self, tmp_path):
-        """Test that saving and loading state preserves optimizer functionality."""
-        optimizer = self._create_optimizer(keep_best_params=False)
-        initial_params = self._create_initial_params()
-        verify_save_load_round_trip(
-            optimizer,
-            initial_params,
-            sphere_cost_fn_population,
-            self.rng,
-            tmp_path,
-            MonteCarloOptimizer.load_state,
-            self.n_params,
-        )
-
-    def test_load_state_raises_file_not_found(self, tmp_path):
-        """Test that load_state() raises CheckpointNotFoundError for missing checkpoint."""
-        verify_load_state_raises_file_not_found(
-            MonteCarloOptimizer.load_state, tmp_path
-        )
-
-    def test_save_state_creates_directory_if_needed(self, tmp_path):
-        """Test that save_state() creates the checkpoint directory if it doesn't exist."""
-        optimizer = self._create_optimizer(keep_best_params=False)
-        initial_params = self._create_initial_params()
-        verify_save_creates_directory_if_needed(
-            optimizer, initial_params, sphere_cost_fn_population, self.rng, tmp_path
-        )
-
-    @pytest.mark.parametrize("keep_best_params", [True, False])
-    def test_compute_new_parameters_preserves_population_size(
-        self, keep_best_params: bool
-    ):
-        """Low level test to ensure sampling logic preserves shapes and bounds."""
-        optimizer = self._create_optimizer(keep_best_params=keep_best_params)
-        population = self._create_initial_params()
-        losses = sphere_cost_fn_population(population)
-        best_indices = np.argsort(losses)[: optimizer.n_best_sets]
-
-        rng = np.random.default_rng(7)
-        new_population = optimizer._compute_new_parameters(
-            population, curr_iteration=1, best_indices=best_indices, rng=rng
-        )
-
-        assert new_population.shape == (optimizer.population_size, self.n_params)
-        assert np.all(new_population >= 0.0)
-        assert np.all(new_population < 2 * np.pi)
-
-        if keep_best_params:
-            np.testing.assert_allclose(
-                new_population[: optimizer.n_best_sets], population[best_indices]
-            )
-
-    def test_optimize_with_completed_iterations_skips_additional_evaluations(self):
-        """Ensure resuming with fewer iterations does not re-evaluate the cost."""
-        optimizer = self._create_optimizer(keep_best_params=False)
-        initial_params = self._create_initial_params()
-
-        eval_counter = {"calls": 0}
-
-        def counting_cost_fn(population: np.ndarray) -> np.ndarray:
-            eval_counter["calls"] += 1
-            return sphere_cost_fn_population(population)
-
-        optimizer.optimize(
-            counting_cost_fn, initial_params, max_iterations=3, rng=self.rng
-        )
-        assert eval_counter["calls"] == 3
-
-        # Resume with a smaller total iteration budget, should finish immediately.
-        result = optimizer.optimize(
-            counting_cost_fn, initial_params, max_iterations=1, rng=self.rng
-        )
-
-        assert eval_counter["calls"] == 3, "No additional evaluations should be made"
-        assert isinstance(result, OptimizeResult)
-        assert result.x.shape == (self.n_params,)
-
-    def test_save_state_without_prior_run_raises(self, tmp_path):
-        """Calling save_state before running optimize should fail clearly."""
-        optimizer = self._create_optimizer(keep_best_params=False)
-        checkpoint_dir = tmp_path / "checkpoint"
-
-        with pytest.raises(RuntimeError, match="optimization has not been run"):
-            optimizer.save_state(str(checkpoint_dir))
-
-    def test_fresh_run_without_initial_params_raises(self):
-        """A fresh run must receive initial_params."""
-        optimizer = self._create_optimizer(keep_best_params=False)
-
-        with pytest.raises(ValueError, match="initial_params is required"):
-            optimizer.optimize(sphere_cost_fn_population, max_iterations=3)
-
-    def test_fresh_run_zero_iterations_raises(self):
-        """Fresh run with max_iterations=0 never evaluates the cost, so no state exists to return."""
-        optimizer = self._create_optimizer(keep_best_params=False)
-        initial_params = self._create_initial_params()
-
-        with pytest.raises(RuntimeError, match="produced no evaluated population"):
-            optimizer.optimize(
-                sphere_cost_fn_population,
-                initial_params,
-                max_iterations=0,
-                rng=self.rng,
-            )
-
-
-class TestPopulationOptimizerCheckpointing:
-    """Shared checkpoint/resume behaviours for population-based optimizers."""
-
-    @pytest.fixture(autouse=True)
-    def setup(self):
-        self.n_params = 4
-        self.rng = np.random.default_rng(1337)
-
-    def _initial_params(self, optimizer: Optimizer) -> np.ndarray:
-        """Create initial parameters that respect the optimizer contract."""
-        shape = (optimizer.n_param_sets, self.n_params)
-        return self.rng.random(shape) * 2 * np.pi
-
-    def test_checkpoint_dir_without_checkpointing(self, checkpointing_optimizer):
-        """Optimization should work normally when no checkpoint_dir is provided."""
-        initial_params = self._initial_params(checkpointing_optimizer)
-
-        result = checkpointing_optimizer.optimize(
-            sphere_cost_fn_population, initial_params, max_iterations=3, rng=self.rng
-        )
-
-        assert isinstance(result, OptimizeResult)
-        assert result.x.shape == (self.n_params,)
-        assert np.isfinite(result.fun)
-
-    def test_resume_with_max_iterations_less_than_completed(
-        self, tmp_path, checkpointing_optimizer
-    ):
-        """Resuming with fewer total iterations should exit immediately."""
-        load_state_func = type(checkpointing_optimizer).load_state
-
-        initial_params = self._initial_params(checkpointing_optimizer)
-
-        checkpointing_optimizer.optimize(
-            sphere_cost_fn_population, initial_params, max_iterations=5, rng=self.rng
-        )
-
-        checkpoint_dir = tmp_path / "checkpoint"
-        checkpointing_optimizer.save_state(str(checkpoint_dir))
-
-        loaded_optimizer = load_state_func(str(checkpoint_dir))
-        result = loaded_optimizer.optimize(
-            sphere_cost_fn_population, initial_params, max_iterations=3, rng=self.rng
-        )
-
-        assert isinstance(result, OptimizeResult)
-        assert result.x.shape == (self.n_params,)
-        assert np.isfinite(result.fun)
-
-    def test_resume_with_different_initial_params(
-        self, tmp_path, checkpointing_optimizer
-    ):
-        """Checkpoints should ignore newly provided initial parameters when resuming."""
-        load_state_func = type(checkpointing_optimizer).load_state
-        initial_params = self._initial_params(checkpointing_optimizer)
-
-        checkpointing_optimizer.optimize(
-            sphere_cost_fn_population, initial_params, max_iterations=2, rng=self.rng
-        )
-
-        checkpoint_dir = tmp_path / "checkpoint"
-        checkpointing_optimizer.save_state(str(checkpoint_dir))
-
-        loaded_optimizer = load_state_func(str(checkpoint_dir))
-        different_initial_params = self._initial_params(checkpointing_optimizer)
-
-        result = loaded_optimizer.optimize(
-            sphere_cost_fn_population,
-            different_initial_params,
-            max_iterations=5,
-            rng=self.rng,
-        )
-
-        assert isinstance(result, OptimizeResult)
-        assert result.x.shape == (self.n_params,)
-        assert np.isfinite(result.fun)
-
-
-class TestPymooOptimizer:
-    """Specific tests for PymooOptimizer features."""
-
-    @pytest.fixture(autouse=True)
-    def setup(self):
-        self.n_params = 4
-        self.rng = np.random.default_rng(42)
-
-    def test_reset_clears_pymoo_state(self):
-        """Test that reset() clears all internal state variables for PymooOptimizer."""
-        optimizer = PymooOptimizer(method=PymooMethod.CMAES, population_size=10)
-        initial_params = self.rng.random((10, self.n_params)) * 2 * np.pi
-
-        # Run optimization to set state
-        optimizer.optimize(
-            sphere_cost_fn_population, initial_params, max_iterations=3, rng=self.rng
-        )
-
-        # Verify state is set
-        assert optimizer._curr_algorithm_obj is not None
-
-        # Reset and verify state is cleared
-        optimizer.reset()
-        assert optimizer._curr_algorithm_obj is None
-
-    def test_save_state_creates_checkpoint_file(self, tmp_path):
-        """Test that save_state() creates the expected checkpoint file."""
-        optimizer = PymooOptimizer(method=PymooMethod.CMAES, population_size=10)
-        initial_params = self.rng.random((10, self.n_params)) * 2 * np.pi
-        verify_save_creates_checkpoint_file(
-            optimizer,
-            initial_params,
-            sphere_cost_fn_population,
-            self.rng,
-            tmp_path,
-            PymooOptimizer.load_state,
-        )
-
-    def test_save_state_without_prior_run_raises(self, tmp_path):
-        """Saving without having run optimize should raise a clear error."""
-        optimizer = PymooOptimizer(method=PymooMethod.CMAES, population_size=5)
-        verify_save_without_prior_run_raises(optimizer, tmp_path)
-
-    def test_save_state_preserves_configuration(self, tmp_path):
-        """Test that save_state() saves optimizer configuration correctly."""
-        optimizer = PymooOptimizer(method=PymooMethod.DE, population_size=15)
-        initial_params = self.rng.random((15, self.n_params)) * 2 * np.pi
-
-        # Run optimization to set state
-        optimizer.optimize(
-            sphere_cost_fn_population, initial_params, max_iterations=2, rng=self.rng
-        )
-
-        checkpoint_dir = str(tmp_path / "checkpoint")
-        optimizer.save_state(checkpoint_dir)
-
-        # Load and verify configuration
-        loaded_optimizer = PymooOptimizer.load_state(checkpoint_dir)
-        assert loaded_optimizer.method == optimizer.method
-        assert loaded_optimizer.population_size == optimizer.population_size
-
-    def test_load_state_restores_optimization_state(self, tmp_path):
-        """Test that load_state() correctly restores optimization state."""
-        optimizer = PymooOptimizer(method=PymooMethod.CMAES, population_size=10)
-        initial_params = self.rng.random((10, self.n_params)) * 2 * np.pi
-
-        # Run optimization to set state
-        optimizer.optimize(
-            sphere_cost_fn_population, initial_params, max_iterations=3, rng=self.rng
-        )
-
-        # Capture algorithm object before saving
-        original_algorithm_obj = optimizer._curr_algorithm_obj
-
-        checkpoint_dir = str(tmp_path / "checkpoint")
-        optimizer.save_state(checkpoint_dir)
-
-        # Load state
-        loaded_optimizer = PymooOptimizer.load_state(checkpoint_dir)
-
-        # Verify algorithm object is restored
-        assert loaded_optimizer._curr_algorithm_obj is not None
-        # The algorithm object should be the same type and have similar state
-        assert type(loaded_optimizer._curr_algorithm_obj) == type(
-            original_algorithm_obj
-        )
-
-    def test_save_load_round_trip(self, tmp_path):
-        """Test that saving and loading state preserves optimizer functionality."""
-        optimizer = PymooOptimizer(method=PymooMethod.CMAES, population_size=10)
-        initial_params = self.rng.random((10, self.n_params)) * 2 * np.pi
-        verify_save_load_round_trip(
-            optimizer,
-            initial_params,
-            sphere_cost_fn_population,
-            self.rng,
-            tmp_path,
-            PymooOptimizer.load_state,
-            self.n_params,
-        )
-
-    def test_load_state_raises_file_not_found(self, tmp_path):
-        """Test that load_state() raises CheckpointNotFoundError for missing checkpoint."""
-        verify_load_state_raises_file_not_found(PymooOptimizer.load_state, tmp_path)
-
-    def test_save_state_creates_directory_if_needed(self, tmp_path):
-        """Test that save_state() creates the checkpoint directory if it doesn't exist."""
-        optimizer = PymooOptimizer(method=PymooMethod.CMAES, population_size=10)
-        initial_params = self.rng.random((10, self.n_params)) * 2 * np.pi
-        verify_save_creates_directory_if_needed(
-            optimizer, initial_params, sphere_cost_fn_population, self.rng, tmp_path
-        )
-
-    def test_resume_extends_iteration_budget(self, tmp_path):
-        """Resuming with a higher max_iterations should continue running iterations."""
-        optimizer = PymooOptimizer(method=PymooMethod.CMAES, population_size=6)
-        initial_params = (
-            self.rng.random((optimizer.n_param_sets, self.n_params)) * 2 * np.pi
-        )
-
-        # Run part of the optimization and checkpoint.
-        optimizer.optimize(
-            sphere_cost_fn_population, initial_params, max_iterations=2, rng=self.rng
-        )
-        checkpoint_dir = str(tmp_path / "checkpoint")
-        optimizer.save_state(checkpoint_dir)
-
-        loaded_optimizer = PymooOptimizer.load_state(checkpoint_dir)
-        result = loaded_optimizer.optimize(
-            sphere_cost_fn_population, max_iterations=4, rng=self.rng
-        )
-
-        assert isinstance(result, OptimizeResult)
-        assert result.nit == 4, "Resume should complete up to the new iteration target"
-
-    def test_resume_when_algorithm_finished(self, tmp_path):
-        """Test resuming when the saved checkpoint was from a completed optimization."""
-        optimizer = PymooOptimizer(method=PymooMethod.CMAES, population_size=10)
-        initial_params = self.rng.random((10, self.n_params)) * 2 * np.pi
-
-        # Run until completion (2 iterations)
-        optimizer.optimize(
-            sphere_cost_fn_population, initial_params, max_iterations=2, rng=self.rng
-        )
-
-        # Verify algorithm finished
-        if optimizer.method == PymooMethod.CMAES:
-            # cma uses countiter
-            assert optimizer._curr_algorithm_obj.countiter == 2
-        else:
-            # n_gen is 1-indexed, so 2 iterations = n_gen 3
-            assert optimizer._curr_algorithm_obj.n_gen == 3
-
-        checkpoint_dir = str(tmp_path / "checkpoint")
-        optimizer.save_state(checkpoint_dir)
-
-        # Resume with more iterations
-        loaded_optimizer = PymooOptimizer.load_state(checkpoint_dir)
-        result = loaded_optimizer.optimize(
-            sphere_cost_fn_population, max_iterations=5, rng=self.rng
-        )
-
-        # Should be able to continue
-        assert isinstance(result, OptimizeResult)
-        assert result.x.shape == (self.n_params,)
-        assert np.isfinite(result.fun)
-
-    def test_n_param_sets_respects_popsize_kwarg(self):
-        """n_param_sets should honor CMAES popsize overrides."""
-        optimizer_with_kwarg = PymooOptimizer(
-            method=PymooMethod.CMAES, population_size=6, popsize=4
-        )
-        assert optimizer_with_kwarg.n_param_sets == 4
-
-        optimizer_default_cmaes = PymooOptimizer(
-            method=PymooMethod.CMAES, population_size=8
-        )
-        assert optimizer_default_cmaes.n_param_sets == 8
-
-        optimizer_de = PymooOptimizer(method=PymooMethod.DE, population_size=11)
-        assert optimizer_de.n_param_sets == 11
-
-    def test_de_save_load_round_trip(self, tmp_path):
-        """Test that DE optimizer can save/load state successfully."""
-        optimizer = PymooOptimizer(method=PymooMethod.DE, population_size=10)
-        initial_params = self.rng.random((10, self.n_params)) * 2 * np.pi
-
-        # Run partial optimization
-        optimizer.optimize(
-            sphere_cost_fn_population, initial_params, max_iterations=2, rng=self.rng
-        )
-
-        checkpoint_dir = str(tmp_path / "checkpoint")
-        optimizer.save_state(checkpoint_dir)
-
-        # Load and continue optimization
-        loaded_optimizer = PymooOptimizer.load_state(checkpoint_dir)
-        result = loaded_optimizer.optimize(
-            sphere_cost_fn_population, max_iterations=5, rng=self.rng
-        )
-
-        # Verify optimization completed successfully
-        assert isinstance(result, OptimizeResult)
-        assert result.x.shape == (self.n_params,)
-        assert np.isfinite(result.fun)
-        assert result.nit == 5, "DE should complete up to the new iteration target"
-
-    def test_de_population_preserved_in_checkpoint(self, tmp_path):
-        """Test that DE's population is properly preserved during checkpointing."""
-        optimizer = PymooOptimizer(method=PymooMethod.DE, population_size=8)
-        initial_params = self.rng.random((8, self.n_params)) * 2 * np.pi
-
-        # Run some iterations
-        optimizer.optimize(
-            sphere_cost_fn_population, initial_params, max_iterations=3, rng=self.rng
-        )
-
-        # Verify internal state exists
-        assert optimizer._curr_algorithm_obj is not None
-        assert hasattr(optimizer._curr_algorithm_obj, "pop")
-        assert optimizer._curr_algorithm_obj.pop is not None
-
-        checkpoint_dir = str(tmp_path / "checkpoint")
-        optimizer.save_state(checkpoint_dir)
-
-        # Load and verify population is restored
-        loaded_optimizer = PymooOptimizer.load_state(checkpoint_dir)
-        assert loaded_optimizer._curr_algorithm_obj is not None
-        assert hasattr(loaded_optimizer._curr_algorithm_obj, "pop")
-        assert loaded_optimizer._curr_algorithm_obj.pop is not None
-
-    def test_cmaes_custom_kwargs_preserved_through_save_load(self, tmp_path):
-        """Test that custom CMAES kwargs survive save/load cycle."""
-        # Create optimizer with custom kwargs
-        optimizer = PymooOptimizer(
-            method=PymooMethod.CMAES, population_size=10, popsize=8, sigma0=0.3
-        )
-        initial_params = self.rng.random((8, self.n_params)) * 2 * np.pi
-
-        # Run optimization
-        optimizer.optimize(
-            sphere_cost_fn_population, initial_params, max_iterations=2, rng=self.rng
-        )
-
-        checkpoint_dir = str(tmp_path / "checkpoint")
-        optimizer.save_state(checkpoint_dir)
-
-        # Load and verify kwargs are preserved
-        loaded_optimizer = PymooOptimizer.load_state(checkpoint_dir)
-        assert loaded_optimizer.population_size == 10
-        assert loaded_optimizer.algorithm_kwargs.get("popsize") == 8
-        assert loaded_optimizer.algorithm_kwargs.get("sigma0") == 0.3
-
-        # Verify it can continue optimization
-        result = loaded_optimizer.optimize(
-            sphere_cost_fn_population, max_iterations=4, rng=self.rng
-        )
-        assert isinstance(result, OptimizeResult)
-        assert result.nit == 4
-
-    def test_optimize_with_zero_max_iterations(self):
-        """Test that optimize with max_iterations=0 returns immediately."""
-        optimizer = PymooOptimizer(method=PymooMethod.CMAES, population_size=5)
-        initial_params = self.rng.random((5, self.n_params)) * 2 * np.pi
-
-        # First run some iterations
-        optimizer.optimize(
-            sphere_cost_fn_population, initial_params, max_iterations=3, rng=self.rng
-        )
-
-        # Now try to resume with max_iterations=0
-        result = optimizer.optimize(
-            sphere_cost_fn_population, max_iterations=0, rng=self.rng
-        )
-
-        # Should return current state without running more iterations
-        assert isinstance(result, OptimizeResult)
-        assert result.nit == 3  # Should still be at iteration 3
-
-    def test_resume_with_max_iterations_equal_to_completed(self):
-        """Test resuming when max_iterations equals already completed iterations."""
-        optimizer = PymooOptimizer(method=PymooMethod.CMAES, population_size=5)
-        initial_params = self.rng.random((5, self.n_params)) * 2 * np.pi
-
-        # Run 5 iterations
-        result1 = optimizer.optimize(
-            sphere_cost_fn_population, initial_params, max_iterations=5, rng=self.rng
-        )
-        assert result1.nit == 5
-
-        # Resume with same max_iterations should not run additional iterations
-        result2 = optimizer.optimize(
-            sphere_cost_fn_population, max_iterations=5, rng=self.rng
-        )
-        assert result2.nit == 5  # Should still be 5, not 10
-
-    def test_fresh_run_without_initial_params_raises(self):
-        """A fresh run must receive initial_params."""
-        optimizer = PymooOptimizer(method=PymooMethod.CMAES, population_size=5)
-
-        with pytest.raises(ValueError, match="initial_params is required"):
-            optimizer.optimize(sphere_cost_fn_population, max_iterations=3)
-
-
-class TestScipyOptimizer:
-    """Specific tests for ScipyOptimizer features."""
-
-    @pytest.fixture(autouse=True)
-    def setup(self):
-        self.n_params = 4
-        self.rng = np.random.default_rng(42)
-
-    def test_reset_is_noop_for_scipy(self):
-        """Test that reset() is a no-op for ScipyOptimizer (no state to clear)."""
-        optimizer = ScipyOptimizer(method=ScipyMethod.L_BFGS_B)
-        initial_params = self.rng.random((1, self.n_params)) * 2 * np.pi
-
-        # Run optimization
-        optimizer.optimize(sphere_cost_fn_single, initial_params, max_iterations=3)
-
-        # Reset should not raise an error and should allow reusing
-        optimizer.reset()
-        result = optimizer.optimize(
-            sphere_cost_fn_single, initial_params, max_iterations=3
-        )
-        assert isinstance(result, OptimizeResult)
-        assert result.x.shape == (self.n_params,)
-
-    def test_save_state_raises_not_implemented(self, tmp_path):
-        """ScipyOptimizer.save_state is not supported."""
-        optimizer = ScipyOptimizer(method=ScipyMethod.L_BFGS_B)
-        with pytest.raises(
-            NotImplementedError, match="ScipyOptimizer does not support"
-        ):
-            optimizer.save_state(str(tmp_path / "checkpoint"))
-
-    def test_load_state_raises_not_implemented(self):
-        """ScipyOptimizer.load_state is not supported."""
-        with pytest.raises(
-            NotImplementedError, match="ScipyOptimizer does not support"
-        ):
-            ScipyOptimizer.load_state("/nonexistent/checkpoint")
-
-    def test_optimize_without_initial_params_raises(self):
-        """ScipyOptimizer cannot resume, so initial_params is always required."""
-        optimizer = ScipyOptimizer(method=ScipyMethod.L_BFGS_B)
-
-        with pytest.raises(ValueError, match="ScipyOptimizer requires initial_params"):
-            optimizer.optimize(sphere_cost_fn_single, max_iterations=3)
-
-    @pytest.mark.parametrize(
-        ("method", "expects_jac"),
-        [
-            (ScipyMethod.L_BFGS_B, True),
-            (ScipyMethod.NELDER_MEAD, False),
-        ],
-    )
-    def test_optimize_passes_jac_only_for_l_bfgs_b(
-        self, method: ScipyMethod, expects_jac: bool, monkeypatch
-    ):
-        """Ensure jacobian callbacks are only forwarded when supported."""
-        recorded = {}
-
-        def fake_minimize(cost_fn, x0, **kwargs):
-            recorded["jac"] = kwargs.get("jac")
-            recorded["method"] = kwargs.get("method")
-            recorded["callback"] = kwargs.get("callback")
-            return OptimizeResult(x=x0, fun=0.0)
-
-        monkeypatch.setattr("divi.qprog.optimizers.minimize", fake_minimize)
-
-        optimizer = ScipyOptimizer(method=method)
-        jac_fn = lambda x: x
-        initial_params = self.rng.random((1, self.n_params)) * 0.1
-
-        optimizer.optimize(
-            sphere_cost_fn_single,
-            initial_params,
-            jac=jac_fn,
-            max_iterations=5,
-        )
-
-        if expects_jac:
-            assert recorded["jac"] is jac_fn
-        else:
-            assert recorded["jac"] is None
-        assert recorded["method"] == method.value
-
-    def test_callback_shapes_are_normalized(self, monkeypatch):
-        """Callbacks should always observe 2D parameters and 1D losses."""
-        captured_shapes: list[tuple[tuple[int, ...], tuple[int, ...]]] = []
-
-        def fake_minimize(cost_fn, x0, **kwargs):
-            callback = kwargs.get("callback")
-            assert callback is not None
-            callback(OptimizeResult(x=np.ones(self.n_params), fun=1.23))
-            return OptimizeResult(x=x0, fun=0.0)
-
-        monkeypatch.setattr("divi.qprog.optimizers.minimize", fake_minimize)
-
-        optimizer = ScipyOptimizer(method=ScipyMethod.NELDER_MEAD)
-        initial_params = self.rng.random((1, self.n_params)) * 0.1
-
-        def tracking_callback(result: OptimizeResult):
-            captured_shapes.append((result.x.shape, result.fun.shape))
-
-        optimizer.optimize(
-            sphere_cost_fn_single,
-            initial_params,
-            max_iterations=3,
-            callback_fn=tracking_callback,
-        )
-
-        assert captured_shapes == [((1, self.n_params), (1,))]
-
-
-class TestCopyOptimizer:
-    """Direct unit tests for ``Optimizer.copy()``."""
-
-    @pytest.fixture(autouse=True)
-    def setup(self):
-        self.n_params = 4
-        self.rng = np.random.default_rng(24)
-
-    def test_copy_monte_carlo_preserves_config_and_resets_state(self):
-        optimizer = MonteCarloOptimizer(
-            population_size=8, n_best_sets=2, keep_best_params=True
-        )
-        initial_params = (
-            self.rng.random((optimizer.population_size, self.n_params)) * 2 * np.pi
-        )
-        optimizer.optimize(
-            sphere_cost_fn_population, initial_params, max_iterations=1, rng=self.rng
-        )
-        assert optimizer._curr_population is not None
-
-        copied = optimizer.copy()
-
-        assert isinstance(copied, MonteCarloOptimizer)
-        assert copied.population_size == optimizer.population_size
-        assert copied.keep_best_params == optimizer.keep_best_params
-        assert copied.n_best_sets == optimizer.n_best_sets
-        assert copied._curr_population is None
-
-    def test_copy_pymoo_preserves_kwargs(self):
-        optimizer = PymooOptimizer(
-            method=PymooMethod.CMAES, population_size=5, popsize=4, sigma=0.5
-        )
-        initial_params = (
-            self.rng.random((optimizer.n_param_sets, self.n_params)) * 2 * np.pi
-        )
-        optimizer.optimize(
-            sphere_cost_fn_population, initial_params, max_iterations=1, rng=self.rng
-        )
-        assert optimizer._curr_algorithm_obj is not None
-
-        copied = optimizer.copy()
-
-        assert isinstance(copied, PymooOptimizer)
-        assert copied.method == optimizer.method
-        assert copied.population_size == optimizer.population_size
-        assert copied.algorithm_kwargs["popsize"] == 4
-        assert copied.algorithm_kwargs["sigma"] == 0.5
-        assert copied._curr_algorithm_obj is None
-
-    def test_copy_scipy_returns_fresh_optimizer(self):
-        optimizer = ScipyOptimizer(method=ScipyMethod.COBYLA)
-        copied = optimizer.copy()
-
-        assert isinstance(copied, ScipyOptimizer)
-        assert copied.method == optimizer.method
-
-    def test_copy_grid_search_returns_fresh_optimizer(self):
-        optimizer = GridSearchOptimizer(param_ranges=[(-1, 1), (-1, 1)], grid_points=5)
-        optimizer.optimize(sphere_cost_fn_population, np.zeros((25, 2)))
-
-        copied = optimizer.copy()
-        assert isinstance(copied, GridSearchOptimizer)
-        assert copied.n_param_sets == 25
-        assert copied._best_params is None  # fresh state
-
-
-class TestGridSearchOptimizer:
-    """Tests for GridSearchOptimizer."""
-
-    # -- Construction --
-
-    def test_param_ranges_creates_cartesian_grid(self):
-        grid = GridSearchOptimizer(param_ranges=[(0, 1), (0, 2)], grid_points=3)
-        assert grid.n_param_sets == 9  # 3 × 3
-
-    def test_param_ranges_1d(self):
-        grid = GridSearchOptimizer(param_ranges=[(-5, 5)], grid_points=100)
-        assert grid.n_param_sets == 100
-
-    def test_explicit_param_grid(self):
-        explicit = np.array([[0.0, 0.0], [1.0, 1.0], [2.0, 2.0]])
-        grid = GridSearchOptimizer(param_grid=explicit)
-        assert grid.n_param_sets == 3
-
-    def test_no_args_raises(self):
-        with pytest.raises(ValueError, match="param_grid or param_ranges"):
-            GridSearchOptimizer()
-
-    def test_1d_param_grid_raises(self):
-        with pytest.raises(ValueError, match="2D"):
-            GridSearchOptimizer(param_grid=np.array([1.0, 2.0, 3.0]))
-
-    # -- Optimization --
-
-    def test_finds_minimum_1d(self):
-        grid = GridSearchOptimizer(param_ranges=[(-5, 5)], grid_points=100)
-        result = grid.optimize(
-            sphere_cost_fn_population, initial_params=np.zeros((100, 1))
-        )
-        assert abs(result.x[0]) < 0.15
-        assert result.fun < 0.03
-        assert result.nit == 1
-
-    def test_finds_minimum_2d(self):
-        grid = GridSearchOptimizer(param_ranges=[(-3, 3), (-3, 3)], grid_points=50)
-        result = grid.optimize(
-            sphere_cost_fn_population, initial_params=np.zeros((1, 2))
-        )
-        assert np.allclose(result.x, [0.0, 0.0], atol=0.15)
-
-    def test_initial_params_ignored(self):
-        """Grid is used regardless of initial_params."""
-        grid = GridSearchOptimizer(param_ranges=[(0, 1)], grid_points=10)
-        result = grid.optimize(
-            sphere_cost_fn_population,
-            initial_params=np.array([[999.0]]),  # should be ignored
-        )
-        assert 0.0 <= result.x[0] <= 1.0
-
-    def test_callback_called(self):
-        grid = GridSearchOptimizer(param_ranges=[(0, 1)], grid_points=5)
-        callbacks = []
-        grid.optimize(
-            sphere_cost_fn_population,
-            initial_params=np.zeros((5, 1)),
-            callback_fn=lambda r: callbacks.append(r),
-        )
-        assert len(callbacks) == 1
-
-    def test_max_iterations_warning(self):
-        grid = GridSearchOptimizer(param_ranges=[(0, 1)], grid_points=5)
-        with pytest.warns(UserWarning, match="max_iterations=10 will be ignored"):
-            grid.optimize(
-                sphere_cost_fn_population,
-                initial_params=np.zeros((5, 1)),
-                max_iterations=10,
-            )
-
-    def test_max_iterations_1_no_warning(self):
-        grid = GridSearchOptimizer(param_ranges=[(0, 1)], grid_points=5)
-        # Should not warn when max_iterations=1 (the default)
-        with warnings.catch_warnings():
-            warnings.simplefilter("error")
-            grid.optimize(
-                sphere_cost_fn_population,
-                initial_params=np.zeros((5, 1)),
-                max_iterations=1,
-            )
-
-    # -- State management --
-
-    def test_reset_clears_state(self):
-        grid = GridSearchOptimizer(param_ranges=[(0, 1)], grid_points=5)
-        grid.optimize(sphere_cost_fn_population, np.zeros((5, 1)))
-        assert grid._best_params is not None
-
-        grid.reset()
-        assert grid._best_params is None
-        assert grid._best_loss is None
-        assert grid._all_losses is None
-
-    def test_get_config(self):
-        grid = GridSearchOptimizer(param_ranges=[(0, 1)], grid_points=10)
-        config = grid.get_config()
-        assert config["type"] == "GridSearchOptimizer"
-        assert config["grid_size"] == 10
-
-    # -- Checkpointing --
-
-    def test_save_and_load_state(self, tmp_path):
-        grid = GridSearchOptimizer(param_ranges=[(-1, 1)], grid_points=20)
-        grid.optimize(sphere_cost_fn_population, np.zeros((20, 1)))
-
-        grid.save_state(tmp_path)
-        loaded = GridSearchOptimizer.load_state(tmp_path)
-
-        assert loaded.n_param_sets == 20
-        np.testing.assert_array_almost_equal(loaded._best_params, grid._best_params)
-        assert loaded._best_loss == pytest.approx(grid._best_loss)
-
-    def test_save_before_optimize_stores_none(self, tmp_path):
-        grid = GridSearchOptimizer(param_ranges=[(0, 1)], grid_points=5)
-        grid.save_state(tmp_path)
-        loaded = GridSearchOptimizer.load_state(tmp_path)
-        assert loaded._best_params is None
-        assert loaded._best_loss is None
